[Feature] Save demo results (#157)
* Delete demo/estimate_smpl_image.py.

* Rename demo/estimate_smpl_video.py -> demo/estimate_smpl.py and update the demo script so that it is compatible with both image-based and video-based methods.

* Save the demo results into human_data.

* Update document for demo/inference.

* Resolve Updated documentation to document how to do inference on a custom dataset. #50.
