+++ docs/getting_started.md
@@ -53,23 +53,26 @@ mmhuman3d
 ```
 
 ## Inference / Demo
+We provide a demo script to estimate SMPL parameters for single-person or multi-person from the input image or video with the bounding box detected by MMDetection or MMTracking. With this demo script, you only need to choose a pre-trained model (we currently only support [HMR](https://github.com/open-mmlab/mmhuman3d/tree/main/configs/hmr/), [SPIN](https://github.com/open-mmlab/mmhuman3d/tree/main/configs/spin/), and [VIBE](https://github.com/open-mmlab/mmhuman3d/tree/main/configs/vibe/), more SOTA methods will be added in the future) from our model zoo and specify a few arguments, and then you can get the estimated results. Moreover, if you specify `--output` and `--show_path`, the demo script will save the estimated results into `human_data` and render the estimated human mesh.
+### Single-person
 
 ```shell
-python demo/estimate_smpl_image.py ${CONFIG_FILE} ${CHECKPOINT} [optional]
+python demo/estimate_smpl.py \
+    ${MMHUMAN3D_CONFIG_FILE} \
+    ${MMHUMAN3D_CHECKPOINT_FILE} \
+    --single_person_demo \
+    --det_config ${MMDET_CONFIG_FILE} \
+    --det_checkpoint ${MMDET_CHECKPOINT_FILE} \
+    --input_path ${VIDEO_PATH_OR_IMG_PATH} \
+    [--show_path ${VIS_OUT_PATH}] \
+    [--output ${RESULT_OUT_PATH}] \
+    [--smooth_type ${SMOOTH_TYPE}] \
+    [--draw_bbox] \
 ```
-
-### Single-person
-Optional arguments include:
-- `--single_person_demo`: flag for single-person inference
-- `--det_config`: MMDetection config
-- `--det_checkpoint`: MMDetection checkpoint
-- `--input_path`: input path
-- `--show_path`: directory to save rendered images or video
-- `--smooth_type`: smoothing mode
 
 Example:
 ```shell
-python demo/estimate_smpl_image.py \
+python demo/estimate_smpl.py \
     configs/hmr/resnet50_hmr_pw3d.py \
     data/checkpoints/resnet50_hmr_pw3d.pth \
     --single_person_demo \
@@ -77,21 +80,26 @@ python demo/estimate_smpl_image.py \
     --det_checkpoint https://download.openmmlab.com/mmdetection/v2.0/faster_rcnn/faster_rcnn_r50_fpn_1x_coco/faster_rcnn_r50_fpn_1x_coco_20200130-047c8118.pth \
     --input_path  demo/resources/single_person_demo.mp4 \
     --show_path vis_results/single_person_demo.mp4 \
-    --smooth_type savgol
+    --output demo_result \
+    --smooth_type savgol \
+    --draw_bbox
 ```
-
-Note that the MMHuman3D checkpoints can be downloaded from the [model zoo](model_zoo.md).
-Here we take HMR (resnet50_hmr_pw3d.pth) as an example.
-
 ### Multi-person
-Optional arguments include:
-- `--multi_person_demo`: flag for multi_person inference
-- `--mmtracking_config`: MMTracking config
-- `--input_path`: input path
-- `--show_path`: directory to save rendered images or video
-- `--smooth_type`: smoothing mode
 
-Example 2: multi-person estimation
+
+```shell
+python demo/estimate_smpl.py \
+    ${MMHUMAN3D_CONFIG_FILE} \
+    ${MMHUMAN3D_CHECKPOINT_FILE} \
+    --multi_person_demo \
+    --tracking_config ${MMTRACKING_CONFIG_FILE} \
+    --input_path ${VIDEO_PATH_OR_IMG_PATH} \
+    [--show_path ${VIS_OUT_PATH}] \
+    [--output ${RESULT_OUT_PATH}] \
+    [--smooth_type ${SMOOTH_TYPE}] \
+    [--draw_bbox]
+```
+Example:
 ```shell
 python demo/estimate_smpl_image.py \
     configs/hmr/resnet50_hmr_pw3d.py \
@@ -100,9 +108,12 @@ python demo/estimate_smpl_image.py \
     --tracking_config demo/mmtracking_cfg/deepsort_faster-rcnn_fpn_4e_mot17-private-half.py \
     --input_path  demo/resources/multi_person_demo.mp4 \
     --show_path vis_results/multi_person_demo.mp4 \
-    --smooth_type savgol
-```
+    --smooth_type savgol \
+    [--draw_bbox]
 
+```
+Note that the MMHuman3D checkpoints can be downloaded from the [model zoo](model_zoo.md).
+Here we take HMR (resnet50_hmr_pw3d.pth) as an example.
 ## Evaluation
 
 We provide pretrained models in the respective method folders in [config](https://github.com/open-mmlab/mmhuman3d/tree/main/configs).
