+++ mmhuman3d/apis/inference.py
@@ -279,12 +279,16 @@ def inference_video_based_model(model,
 
     for idx in range(len(feature_sequences)):
         mesh_result = dict()
-        # mesh_result['track_id'] = feature_sequences[idx]['track_id']
         mesh_result['camera'] = results['camera'][idx, target_idx]
         mesh_result['smpl_pose'] = results['smpl_pose'][idx, target_idx]
         mesh_result['smpl_beta'] = results['smpl_beta'][idx, target_idx]
         mesh_result['vertices'] = results['vertices'][idx, target_idx]
         mesh_result['keypoints_3d'] = results['keypoints_3d'][idx, target_idx]
+        mesh_result['bbox'] = extracted_results[target_idx][idx]['bbox']
+        # 'track_id' is not included in results generated by mmdet
+        if 'track_id' in extracted_results[target_idx][idx].keys():
+            mesh_result['track_id'] = extracted_results[target_idx][idx][
+                'track_id']
         mesh_results.append(mesh_result)
     return mesh_results
